Updating mps scores and total score in Autoevaluation instances
