+++ mm_evaluation/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class MmEvaluationConfig(AppConfig):
     name = 'mm_evaluation'
+
+    def ready(self):
+        import mm_evaluation.signals
